+++ src/services/loade-component.js
